// user_service/src/generated/user/v1/user.ts
export enum ResumeStatus {
  RESUME_STATUS_UNSPECIFIED = 0,
  RESUME_STATUS_DRAFT = 1,
  RESUME_STATUS_PUBLISHED = 2,
  RESUME_STATUS_ARCHIVED = 3,
  UNRECOGNIZED = -1,
}

export enum ResumeVisibility {
  RESUME_VISIBILITY_UNSPECIFIED = 0,
  RESUME_VISIBILITY_PUBLIC = 1,
  RESUME_VISIBILITY_REGISTERED_EMPLOYERS = 2,
  RESUME_VISIBILITY_PRIVATE = 3,
  UNRECOGNIZED = -1,
}

export enum SearchStatus {
  SEARCH_STATUS_UNSPECIFIED = 0,
  SEARCH_STATUS_ACTIVE_SEARCH = 1,
  SEARCH_STATUS_OPEN_TO_OFFERS = 2,
  SEARCH_STATUS_NOT_SEARCHING = 3,
  SEARCH_STATUS_FOUND_JOB = 4,
  UNRECOGNIZED = -1,
}

export enum BusinessTripReadiness {
  BUSINESS_TRIP_READINESS_UNSPECIFIED = 0,
  BUSINESS_TRIP_READINESS_NEVER = 1,
  BUSINESS_TRIP_READINESS_SOMETIMES = 2,
  BUSINESS_TRIP_READINESS_READY = 3,
  UNRECOGNIZED = -1,
}

export enum EmploymentType {
  EMPLOYMENT_TYPE_UNSPECIFIED = 0,
  EMPLOYMENT_TYPE_FULL_TIME = 1,
  EMPLOYMENT_TYPE_PART_TIME = 2,
  EMPLOYMENT_TYPE_PROJECT = 3,
  EMPLOYMENT_TYPE_INTERNSHIP = 4,
  EMPLOYMENT_TYPE_TEMPORARY = 5,
  EMPLOYMENT_TYPE_VOLUNTEER = 6,
  UNRECOGNIZED = -1,
}

export enum WorkSchedule {
  WORK_SCHEDULE_UNSPECIFIED = 0,
  WORK_SCHEDULE_FULL_DAY = 1,
  WORK_SCHEDULE_SHIFT = 2,
  WORK_SCHEDULE_FLEXIBLE = 3,
  WORK_SCHEDULE_ROTATION = 4,
  UNRECOGNIZED = -1,
}

export enum WorkFormat {
  WORK_FORMAT_UNSPECIFIED = 0,
  WORK_FORMAT_ONSITE = 1,
  WORK_FORMAT_REMOTE = 2,
  WORK_FORMAT_HYBRID = 3,
  WORK_FORMAT_FIELD = 4,
  UNRECOGNIZED = -1,
}
// ...
export interface ResumeDraft {
  title: string;
  about?: string | undefined;
  city?: string | undefined;
  visibility?: ResumeVisibility | undefined;
  search_status?: SearchStatus | undefined;
  business_trip_readiness?: BusinessTripReadiness | undefined;
  salary_amount?: number | undefined;
  salary_currency?: string | undefined;
  employment_types: EmploymentType[];
  work_schedules: WorkSchedule[];
  work_formats: WorkFormat[];
  education_ids: string[];
  work_experience_ids: string[];
  skill_ids: number[];
  professional_role_ids: number[];
  certificates: CertificateInput[];
}

export interface CertificateInput {
  title: string;
  issuer?: string | undefined;
  url?: string | undefined;
  achieved_at?: string | undefined;
}